use useBlob + abstract to FormTemplateLayout
Changes to blob:
- includes size as a string for form display on upload page
- takes a file rather than a change event because dropbox processes change events and gives a list of accepted files
