+++ apps/web/src/components/createFormTemplate/FormTemplateButtons.tsx
@@ -8,7 +8,7 @@ import { useRouter } from 'next/router';
  * @param backLink page router will push on click of 'back'
  * @param disabled whether the 'save & continue' button should be activated
  */
-export const FormTemButtons = ({
+export const FormTemplateButtons = ({
   deleteFunction,
   submitLink,
   backLink,
